+++ CHANGELOG.md
@@ -2,7 +2,9 @@ Below are the noteworthy changes from each release.
 A more detailed list of changes is available in the corresponding milestones for each release in the Github issue tracker (https://github.com/googlefonts/fontbakery/milestones?state=closed).
 
 ## Upcoming release: 0.12.9 (2024-Jul-??)
-  - ...
+### Changes to existing checks
+#### On the Universal profile
+  - **[com.google.fonts/check/gsub/smallcaps_before_ligatures]:** Skip check if font lacks GSUB table or font is missing either liga or smcp features. (PR #4787)
 
 
 ## 0.12.8 (2024-Jul-05)